Roll back 64-bit fixes because they are potentially destabilizing and didn't fix the 64-bit issues anyhow
git-svn-id: svn+ssh://svn.code.sf.net/p/turbovnc/code/trunk@811 799e4f7b-5fd2-41f6-823c-2ecc41bc7f0b

+++ vnc_unixsrc/Xvnc/programs/Xserver/hw/vnc/sprite.c
@@ -194,7 +194,7 @@ static GCOps rfbSpriteGCOps = {
     (rfbScreen.cursorIsDrawn &&					    \
         (pScreenPriv->pCacheWin == pWin ?			    \
 	    pScreenPriv->isInCacheWin : (			    \
-	    ((long) (pScreenPriv->pCacheWin = (pWin))) ,		    \
+	    ((int) (pScreenPriv->pCacheWin = (pWin))) ,		    \
 	    (pScreenPriv->isInCacheWin =			    \
 		(pWin)->drawable.x < pScreenPriv->saved.x2 &&	    \
 		pScreenPriv->saved.x1 < (pWin)->drawable.x +	    \
